move semantic components to folder and add spinner to detail
