+++ portal/code/tractus-x-portal/src/components/semanticmodels/new.tsx
